Teach app view about A2 embedded server and streamline run config.
