+++ packages/unified-server/package.json
@@ -169,15 +169,16 @@
     "wireit": "^0.15.0-pre.2"
   },
   "dependencies": {
-    "@google-labs/breadboard": "0.33.0",
+    "@breadboard-ai/a2": "^0.3.0",
     "@breadboard-ai/board-server": "0.13.0",
     "@breadboard-ai/board-server-management": "1.20.1",
-    "@breadboard-ai/connection-server": "0.5.1",
     "@breadboard-ai/connection-client": "0.2.0",
+    "@breadboard-ai/connection-server": "0.5.1",
     "@breadboard-ai/data-store": "0.3.2",
     "@breadboard-ai/remote-board-server": "1.21.0",
     "@breadboard-ai/shared-ui": "1.24.0",
     "@breadboard-ai/visual-editor": "1.27.0",
+    "@google-labs/breadboard": "0.33.0",
     "@lit/context": "^1.1.4",
     "express": "^4.21.2",
     "vite": "^6.2.6",
